Changed type from decimal to int

+++ MurderWithFriendsAPI/Controllers/DamageController.cs
@@ -29,7 +29,7 @@ namespace MurderWithFriendsAPI.Controllers
 
         // GET: api/Damage/5
         [HttpGet("{attackerid}/{defenderid}", Name = "Get")]
-        public decimal Get(long attackerId, long defenderId)
+        public int Get(long attackerId, long defenderId)
         {
 			return _damageService.CalculateDamage(attackerId, defenderId);
         }
